pyproject.toml
[tool.poetry]
name = "nonebot_plugin_shindan"
version = "0.1.3"
description = "Nonebot2 plugin for using ShindanMaker"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-alpha.15"
nonebot-adapter-cqhttp = "^2.0.0-alpha.15"
httpx = "^0.19.0"
lxml = "^4.6.4"
playwright = "^1.9.2"

[tool.poetry.dev-dependencies]
nb-cli = "^0.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
